refactor: simplify operation trait based on parser-api
